requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
python-dotenv
pytest
httpx
python-multipart
pydantic-settings
bcrypt
